requirements.txt
Django==6.0.3
djangorestframework==3.16.1
drf-yasg==1.21.5
djangorestframework_simplejwt==5.5.1
django-filter==25.2
python-dotenv==1.2.2